seasonal flu resolution
